+++ tests/integration/test_installation.py
@@ -6,7 +6,6 @@ from collections.abc import Callable
 from dataclasses import replace
 from datetime import timedelta
 
-import databricks.sdk.errors
 import pytest  # pylint: disable=wrong-import-order
 from databricks.labs.blueprint.installation import Installation
 from databricks.labs.blueprint.installer import InstallState, RawState
@@ -18,9 +17,11 @@ from databricks.sdk.retries import retried
 from databricks.sdk.service import compute, sql
 from databricks.sdk.service.iam import PermissionLevel
 
+import databricks
 from databricks.labs.ucx.config import WorkspaceConfig
 from databricks.labs.ucx.hive_metastore.mapping import Rule
 from databricks.labs.ucx.install import WorkspaceInstallation, WorkspaceInstaller
+from databricks.labs.ucx.installer.workflows import WorkflowsInstallation
 from databricks.labs.ucx.workspace_access import redash
 from databricks.labs.ucx.workspace_access.generic import (
     GenericPermissionsSupport,
@@ -94,14 +95,16 @@ def new_installation(ws, sql_backend, env_or_skip, inventory_schema):
         # TODO: see if we want to move building wheel as a context manager for yield factory,
         # so that we can shave off couple of seconds and build wheel only once per session
         # instead of every test
+        workflows_installation = WorkflowsInstallation(
+            workspace_config, installation, ws, product_info.wheels(ws), prompts, product_info, timedelta(minutes=2)
+        )
         workspace_installation = WorkspaceInstallation(
             workspace_config,
             installation,
             sql_backend,
-            product_info.wheels(ws),
             ws,
+            workflows_installation,
             prompts,
-            timedelta(minutes=2),
             product_info,
         )
         workspace_installation.run()
